web/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=web/requirements-dev.txt requirements/dev.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-stubs
attrs==24.2.0
    # via
    #   flake8-bugbear
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via
    #   -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
    #   metadata-parser
black==24.10.0
    # via -r requirements/dev.in
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via black
colorama==0.4.6
    # via click
coverage==7.6.4
    # via -r requirements/dev.in
cryptography==43.0.3
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-rest-auth[with-social]==6.0.0
    # via -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
django==5.0
    # via
    #   -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
    #   dj-rest-auth
    #   django-allauth
    #   django-debug-toolbar
    #   django-filter
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth==0.61.1
    # via dj-rest-auth
django-debug-toolbar==4.4.6
    # via -r requirements/dev.in
django-dotenv==1.4.2
    # via -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
django-filter==24.3
    # via -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
django-stubs==5.1.1
    # via
    #   -r requirements/dev.in
    #   djangorestframework-stubs
django-stubs-ext==5.1.1
    # via django-stubs
djangorestframework==3.15.0
    # via
    #   -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
djangorestframework-stubs==3.15.1
    # via -r requirements/dev.in
drf-spectacular==0.27.2
    # via -r requirements/dev.in
factory-boy==3.3.1
    # via -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
faker==30.8.2
    # via
    #   -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
    #   factory-boy
flake8==7.1.1
    # via
    #   -r requirements/dev.in
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-commas
    #   pep8-naming
flake8-bugbear==24.8.19
    # via -r requirements/dev.in
flake8-builtins==2.5.0
    # via -r requirements/dev.in
flake8-commas==4.0.0
    # via -r requirements/dev.in
flake8-variables-names==0.0.6
    # via -r requirements/dev.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
isort==5.13.2
    # via -r requirements/dev.in
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
mccabe==0.7.0
    # via flake8
metadata-parser==0.12.1
    # via -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
mypy==1.13.0
    # via -r requirements/dev.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.1
    # via black
pathspec==0.12.1
    # via black
pep8-naming==0.14.1
    # via -r requirements/dev.in
platformdirs==4.3.6
    # via black
psycopg2-binary==2.9.10
    # via -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via cffi
pyflakes==3.2.0
    # via flake8
pyjwt[crypto]==2.9.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via faker
python3-openid==3.2.0
    # via django-allauth
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r C:\Shizuo\Repos\backend\drf\link-storage\requirements\base.in
    #   django-allauth
    #   djangorestframework-stubs
    #   metadata-parser
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via django-allauth
requests-toolbelt==1.0.0
    # via metadata-parser
rpds-py==0.20.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
soupsieve==2.6
    # via beautifulsoup4
sqlparse==0.5.1
    # via
    #   django
    #   django-debug-toolbar
types-pyyaml==6.0.12.20240917
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.32.0.20241016
    # via djangorestframework-stubs
typing-extensions==4.12.2
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   faker
    #   mypy
tzdata==2024.2
    # via django
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via
    #   requests
    #   types-requests
